# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkyHostSolutions API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessMessage {
  message: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface ClientRegisterInput {
  name: string;
  email: string;
  password: string;
  phone: string;
  /** @nullable */
  company?: string | null;
}

export interface AuthResponse {
  id: number;
  name: string;
  email: string;
  phone?: string;
  /** @nullable */
  company?: string | null;
  status: string;
}

export interface AdminAuthResponse {
  id: number;
  username: string;
  role: string;
}

export interface ClientUser {
  id: number;
  name: string;
  email: string;
  phone?: string;
  /** @nullable */
  company?: string | null;
  status: string;
  createdAt: string;
}

export interface ClientUpdate {
  name?: string;
  phone?: string;
  /** @nullable */
  company?: string | null;
  status?: string;
}

export interface Service {
  id: number;
  name: string;
  slug: string;
  category: string;
  shortDescription: string;
  description?: string;
  features?: string[];
  priceUsd: number;
  priceEur: number;
  priceInr: number;
  imageUrl: string;
  enabled: boolean;
}

export interface ServiceDetail {
  service: Service;
  relatedServices: Service[];
}

export interface ServiceInput {
  name: string;
  slug: string;
  category: string;
  shortDescription: string;
  description: string;
  features?: string[];
  priceUsd: number;
  priceEur: number;
  priceInr: number;
  imageUrl: string;
  enabled?: boolean;
}

export interface ServiceUpdate {
  name?: string;
  shortDescription?: string;
  description?: string;
  features?: string[];
  priceUsd?: number;
  priceEur?: number;
  priceInr?: number;
  imageUrl?: string;
  enabled?: boolean;
}

export interface Testimonial {
  id: number;
  name: string;
  company: string;
  rating: number;
  review: string;
  avatarUrl: string;
}

export interface ContactInput {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
}

export interface Order {
  id: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  serviceId: number;
  /** @nullable */
  serviceName?: string | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  amount: number;
  currency: string;
  createdAt: string;
  updatedAt?: string;
}

export interface OrderInput {
  clientId: number;
  serviceId: number;
  /** @nullable */
  notes?: string | null;
  amount: number;
  currency: string;
}

export interface OrderUpdate {
  status?: string;
  /** @nullable */
  notes?: string | null;
}

export interface Invoice {
  id: number;
  invoiceNumber?: string;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  orderId: number | null;
  amount: number;
  currency: string;
  status: string;
  dueDate: string;
  /** @nullable */
  paidAt?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface InvoiceInput {
  clientId: number;
  /** @nullable */
  orderId?: number | null;
  amount: number;
  currency: string;
  dueDate: string;
  /** @nullable */
  notes?: string | null;
}

export interface InvoiceUpdate {
  status?: string;
  dueDate?: string;
  /** @nullable */
  paidAt?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Ticket {
  id: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  subject: string;
  status: string;
  priority: string;
  createdAt: string;
  updatedAt?: string;
}

export interface TicketInput {
  subject: string;
  message: string;
  priority: string;
}

export interface TicketUpdate {
  status?: string;
  priority?: string;
}

export interface TicketReply {
  id: number;
  ticketId: number;
  authorName?: string;
  message: string;
  isAdmin: boolean;
  createdAt: string;
}

export interface TicketReplyInput {
  message: string;
}

export interface TicketDetail {
  ticket: Ticket;
  replies: TicketReply[];
}

export interface ClientDashboard {
  totalOrders: number;
  activeOrders: number;
  openTickets: number;
  unpaidInvoices: number;
  recentOrders: Order[];
  recentTickets: Ticket[];
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface MonthRevenue {
  month: string;
  revenue: number;
}

export interface AdminDashboard {
  totalClients: number;
  totalOrders: number;
  activeOrders: number;
  openTickets: number;
  totalRevenue: number;
  recentOrders: Order[];
  recentTickets: Ticket[];
  ordersByStatus: StatusCount[];
  revenueByMonth: MonthRevenue[];
}