# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SkyHostSolutions API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a single service with related services
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDetail"
        "404":
          description: Not found

  # --- Testimonials ---
...
  /auth/register:
    post:
      operationId: registerClient
      tags: [auth]
      summary: Register a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientRegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: loginClient
      tags: [auth]
      summary: Client login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/admin/login:
    post:
      operationId: loginAdmin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Admin logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized

  # --- Client Portal ---
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    ClientRegisterInput:
      type: object
      required: [name, email, password, phone]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        company:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [id, name, email, status]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: ["string", "null"]
        status:
          type: string
    AdminAuthResponse:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
...
    Service:
      type: object
      required: [id, name, slug, category, shortDescription, priceUsd, priceEur, priceInr, imageUrl, enabled]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        shortDescription:
          type: string
        description:
          type: string
        features:
          type: array
          items:
            type: string
        priceUsd:
          type: number
        priceEur:
          type: number
        priceInr:
          type: number
        imageUrl:
          type: string
        enabled:
          type: boolean
    ServiceDetail:
      type: object
      required: [service, relatedServices]
      properties:
        service:
          $ref: "#/components/schemas/Service"
        relatedServices:
          type: array
          items:
            $ref: "#/components/schemas/Service"
...