# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkyHostSolutions API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  slug: string;
  category: string;
  shortDescription: string;
  description?: string;
  features?: string[];
  priceUsd: number;
  priceEur: number;
  priceInr: number;
  imageUrl: string;
  enabled: boolean;
}